RIA-5804 Adapter layer for document management service (unverified)
