+++ app/controllers/upload-evidence/upload-evidence-controller.ts
@@ -1,9 +1,10 @@
 import { NextFunction, Request, Response } from 'express';
 import i18n from '../../../locale/en.json';
-import { documentIdToDocStoreUrl, DocumentManagementService } from '../../service/document-management-service';
+import { DocumentManagementService } from '../../service/document-management-service';
 import UpdateAppealService from '../../service/update-appeal-service';
 import { getConditionalRedirectUrl } from '../../utils/url-utils';
 import { createStructuredError, yesOrNoRequiredValidation } from '../../utils/validations/fields-validations';
+import { documentIdToDocStoreUrl } from '../../utils/utils';
 
 export interface EvidenceUploadConfig {
   evidenceYesNoPath: string;
